requirements.txt
chromadb
openai
python-dotenv
fastapi
uvicorn
pydantic
httpx
fastmcp
